Can't even remember what I've fixed this time

+++ metro.pde
@@ -1,6 +1,6 @@
 class metro{
   int tempo;
-  int prevTick;
+  long prevTick;
   int beattime;
   int count;
   metro(int tempoarg,int divisions){
